+++ Core/Src/NTC_Temp_Reading.c
@@ -2,8 +2,8 @@
 
 float NTC_Read(uint16_t analog_val, float R0)
 {
-    double resistance = (analog_val * 10000.0f)/(65536.0f - analog_val);
-    double temp_K = resistance/R0;
+    float resistance = (analog_val * 10000.0f)/(65536.0f - analog_val);
+    float temp_K = resistance/R0;
 
     temp_K = log(temp_K);
     temp_K /= 4000.0f;
